:tractor: split server and client
