+++ src/outlier/biz/bizservice.py
@@ -208,7 +208,7 @@ def bizclnt(state:State,
                 bizservice.send(bizreq)
                 return bizreq
             except Exception as e:
-                logging.debug(e)
+                print(e)
                 print(wrapper.help)
                 return BizRequest()
        
@@ -293,271 +293,12 @@ class ClientBizService(BizService):
 
 
 
-@dataclass
-class ChatMessage:
-    name:       str
-    ts:         float
-    content:    str
 
-    def getattrs(self):
-        return {'name': self.name, 'timestamp': self.ts, 'content': self.content}
-    
-    @staticmethod
-    def parse(cm: dict):
-        return ChatMessage(cm['name'], cm['timestamp'], cm['content'])
-    
-    @staticmethod
-    def get_round_time(timestamp: float) -> str:
-        """This is to convert timestamp gap to a human-readable time gap
-        such as `1 days ago`, `1 hours ago` etc.
-
-        Args:
-            timestamp (float): target timestamp
-
-        Returns:
-            str: time gap
-        """
-        curr = datetime.now()
-        happening = datetime.fromtimestamp(timestamp)
-        delta = curr - happening
-        if (delta.days > 0):
-            return f"{delta.days} days ago"
-        elif delta.seconds // 3600 > 0:
-            return f"{delta.seconds // 3600} hours ago"
-        elif delta.seconds // 60 > 0:
-            return f"{delta.seconds // 60} minutes ago"
-        else:
-            return "just now"
-    
-    def format(self):
-        return f"{self.name} -- {ChatMessage.get_round_time(self.ts)}: {self.content}\n"
-    
-
-class Room:
-    index = 0
-    def __init__(self, name) -> None:
-        self.history: List[ChatMessage] = list()
-        self.group:   int               = Room.index
-        self.passwd:  str               = None
-        self.name:    str               = name
-        self.connects:int               = 0
-        Room.index += 1
-        
-    
-    def __repr__(self) -> str:
-        l = ("PSWD" if self.passwd else "FREE")
-        return f"{'room:'+str(self.name):10}\t{l}\t{self.connects}\t{len(self.history)}"
-    
-    @staticmethod
-    def header():
-        return f"{'name':10}\t{'LOCK'}\tpeople\tmsg"
-    
-
-    
-class Server(ServerBizService):
-    def __init__(self, sessservice: SessionService, **kwargs) -> None:
-        super().__init__(sessservice, **kwargs)
-        self.rooms: List[Room] = [Room("R1")]
-        
-
-    @bizserv()
-    def chat(self, user: User, bizreq: BizRequest) -> BizResponse:
-        searchroom = None
-        for room in self.rooms:
-            if room.group == user.sess.group:
-                logging.debug(f"find Room {room}")
-                searchroom = room   
-                break
-            
-        if searchroom is not None:            
-            searchroom.history.append(ChatMessage.parse(bizreq.param))
-            return BizResponse(bizreq.param, bizreq.param, group=room.group, inc=True)
-        else:
-            return BizResponse("error")
-    
-    
-    @bizserv()
-    def connectuser(self, user: User, bizreq: BizRequest) -> BizResponse:
-        user.name = bizreq.param
-        return BizResponse(f"Welcome {bizreq.param}")
-    
-    @bizserv()
-    def infohall(self, user: User, bizreq: BizRequest) -> BizResponse:
-        def getinfo():
-            ret = ""
-            ret += Room.header() + "\n"
-            for room in self.rooms:
-                ret += str(room) + "\n"
-            return ret
-        info = getinfo()
-        return BizResponse(f"{info}", None)
-    
-    @bizserv()
-    def enterroom(self, user: User, bizreq: BizRequest) -> BizResponse:
-        name = bizreq.param
-        for room in self.rooms:
-            if room.name == name:
-                logging.debug(f"enter room {room.name}")
-                user.sess.group = room.group
-                room.connects += 1
-        return BizResponse(f"enter room {room.name}", f"{user.name} entered", group=room.group, inc=False)
-    
-    @bizserv(rc=Ops.Rmv)
-    def leftroom(self, user: User, bizreq: BizRequest) -> BizResponse:
-        group = -1
-        for room in self.rooms:
-            if room.group == user.sess.group:
-                logging.debug(f"left room {room.name}")
-                group = room.group
-                user.sess.group = -1
-                room.connects -= 1
-                break
-        if group != -1:
-            return BizResponse(f"leave room {room.name}", f"{user.name} left", group=group)
-        else:
-            return BizResponse()
-        
-    @bizserv()
-    def roominfo(self, user: User, bizreq: BizRequest) -> BizResponse:
-        def getinfo():
-            searchroom = None
-            for room in self.rooms:
-                if room.group == user.sess.group:
-                    logging.debug(f"find Room {room}")
-                    searchroom = room   
-                    break
-                
-            if searchroom is not None:
-                ret = f"Room: {searchroom.name}\n"
-                for u in self.users.values():
-                    if u.sess.group == searchroom.group:
-                        ret += f"Connecting: {u.name}\tat {u.sess.conn.addr}\n"
-                ret += f"room history message size: {len(searchroom.history)}\n"
-                return ret
-            else:
-                return "error"
-        ret = getinfo()
-        return BizResponse(ret)
             
         
     
     
     
     
-class Client(ClientBizService):
-    def __init__(self, sessservice: SessionService, **kwargs) -> None:
-        super().__init__(sessservice, **kwargs)
-        
-        self.chatblock = False
-        self.buffer: List[ChatMessage] = list()
-        
-    def show_msg(self, package):
-        print(f"{package.get_field('param')}")
-        
-    def show_chat(self, package):
-        cm = package.get_field('param')
-        cm = ChatMessage.parse(cm)
-        if self.chatblock:
-            self.buffer.append(cm)
-        else:
-            print(cm.format())
-        
-    
-    def connected(self, package):
-        self.atstate = State.Hall
-        self.show_msg(package)
-    
-    @bizclnt(state=State.IDLE, bindto=Server.connectuser, recall=connected)
-    def connect(self, inputs=None, *args, **kwargs) -> BizRequest:
-        return BizRequest(param="Alice")
-    
-    
-    
-    @bizclnt(state=State.Room, invokeptn="^(?!\$).+", bindto=Server.chat, recall=show_chat)
-    def chat(self, inputs=None, *args, **kwargs) -> BizRequest:
-        from datetime import datetime
-        cm = ChatMessage("Alice", datetime.now().timestamp(), inputs)
-        return BizRequest(param=cm.getattrs())
-    
-    
-    
-    @bizclnt(state=State.Hall, invoke="$info", bindto=Server.infohall, recall=show_msg, 
-             desc="To glance at the details of the server and client configuartions")
-    def info(self, inputs=None, *args, **kwargs) -> BizRequest:
-        return BizRequest()
-    
-    
-    
-    @bizclnt(state=State.Hall, invoke="$exit", 
-             desc="To disconnect with the server")
-    def quit(self, inputs=None, *args, **kwargs):
-        self.close()
-        return BizRequest()
-        
-    
-    @bizclnt(state=State.Hall, invoke="$help")
-    def help(self, inputs=None, *args, **kwargs) -> BizRequest:
-        return BizRequest()
-    
-    
-    
-    
-    def entered(self, package):
-        self.atstate = State.Room
-        self.show_msg(package)
-    
-    @bizclnt(state=State.Hall, invoke="$room", bindto=Server.enterroom, recall=entered,
-            descparams="[room name]",
-            desc="To enter one of the room or create one by specified name or your username")
-    def enter(self, inputs=None, *args, **kwargs) -> BizRequest:
-        if len(args) == 1:
-            return BizRequest(param=args[0])
-        else:
-            raise Exception("Parameter Error")
-    
-    
-    
-    def left(self, package):
-        self.atstate = State.Hall
-        self.show_msg(package)
 
-    @bizclnt(state=State.Room, invoke="$leave", bindto=Server.leftroom, recall=left)
-    def leave(self, inputs=None, *args, **kwargs) -> BizRequest:
-        return BizRequest()
-    
-    
-    
-    @bizclnt(state=State.Room, invoke="$info", bindto=Server.roominfo, recall=show_msg,
-            desc="To glance at the details of the room")
-    def roominfo(self, inputs=None, *args, **kwargs) -> BizRequest:
-        return BizRequest()
-    
-    
-    @bizclnt(state=State.Room, invoke="$clear",
-            desc="To clear the terminal")
-    def clear(self, inputs=None, *args, **kwargs) -> BizRequest:
-        import subprocess
-        subprocess.call('reset')
-        return BizRequest()
-    
-    
-    @bizclnt(state=State.Room, invoke="$",
-            desc="Switch comming message block")
-    def blockswitch(self, inputs=None, *args, **kwargs) -> BizRequest:
-        if self.chatblock:
-            self.chatblock = False
-            for cm in self.buffer:
-                print(cm.format())
-        else:
-            self.chatblock = True
-        return BizRequest()
         
-        
-    
-    def start(self):
-        super().start()
-        time.sleep(0.5)
-        self.connect()
-        while True:
-            a = input()
-            self.process_input(a)
